[change]: code cleanup + codefactor complexity
